Periodically invoking System.gc() to hint to JVM that a garbage collection should happen. Removing BigInteger simtime in favor of long simTime; BigInteger is just too bulky to use as a timestamp.

+++ trunk/src/dars/SimulationTimeKeeper.java
@@ -3,5 +3,5 @@ package dars;
 import java.math.BigInteger;
 
 public interface SimulationTimeKeeper {
-  BigInteger getTime();
+  long getTime();
 }
